Move `MainQueueDispatchDecorator` to a separate file
